Fixed up a few things in the UI
We were writed up to always run the test, and I realized that it was
just as good to assume a test if the cyphertext was missing. So I added
that to the solver, and it's all looking good. I also corrected the
getting and setting of the NSPopUpButton as the stringValue: worked, but
we really could do so much better if we did this based on the Title. Now
it's looking very good indeed.