Updated the README
We saw a flurry of activity on these READMEs as it was really time to
have a little fun with the presentation and Xcode is kinda fun when you
don't have to build monster projects in it. This is just a little bit of
help for the reader so they know what's happening and why.