Fixed up for Swift 5.0
We updated to Xcode 10.2 and that means that Swift 5.0 is the new
default, and so we had a few changes to make on the string handling and
I have to say I don't even begin to understand the logic that went
behind this. Horrible... just bad - even for the complexities of
Strings, it's bad.