Updated the 'pattern' method
The original implementation of 'pattern' on  String wasn't as nice as I
really wanted it to be, and this new one uses a simply array conversion
of the UTF8 string and we're a lot simpler here. This doesn't make any
difference in the execution speed, but it does make the code a little
cleaner.