Realized we needed to lowercase
While working on the ObjC version of the code, I was surprised that I
needed to lowercase the pattern source so that something like 'Adam'
would be patterned properly as 'abac' and not treat the capitalization
as a different letter than the third letter. I hadn't run into this in
the testing, but it was something that certainly needed to be done.